+++ apps/website/CHANGELOG.md
@@ -1,5 +1,19 @@
 # website
 
+## 0.0.39
+
+### Patch Changes
+
+- Updated dependencies
+  - @visactor/vseed@0.0.39
+
+## 0.0.38
+
+### Patch Changes
+
+- Updated dependencies
+  - @visactor/vseed@0.0.38
+
 ## 0.0.37
 
 ### Patch Changes